requirements.txt
pypdf
langchain~=0.0.284
torch~=2.0.1
accelerate
bitsandbytes
transformers~=4.33.1
sentence_transformers
faiss_cpu
chainlit~=0.6.402
scipy~=1.11.2
flask~=2.3.3
flask_cors
python-dotenv~=1.0.0
pymongo~=4.5.0
flask-PyMongo~=2.3.0
diffusers~=0.21.1
requests~=2.31.0
Pillow~=9.5.0
cohere~=4.27